fix: types ownership
