+++ packages/design-system-react/src/components/TextField/TextField.stories.tsx
@@ -1,4 +1,3 @@
-import { TextFieldSize, TextFieldType } from '@metamask/design-system-shared';
 import type { Meta, StoryObj } from '@storybook/react-vite';
 import React, { useEffect, useState } from 'react';
 
@@ -6,6 +5,7 @@ import { Text } from '../Text';
 
 import README from './README.mdx';
 import { TextField } from './TextField';
+import { TextFieldSize, TextFieldType } from './TextField.types';
 import type { TextFieldProps } from './TextField.types';
 
 function ControlledTextField(props: TextFieldProps) {
