+++ packages/design-system-react/src/components/TextField/TextField.test.tsx
@@ -1,8 +1,8 @@
-import { TextFieldSize, TextFieldType } from '@metamask/design-system-shared';
 import { fireEvent, render, screen } from '@testing-library/react';
 import React, { createRef } from 'react';
 
 import { TextField } from './TextField';
+import { TextFieldSize, TextFieldType } from './TextField.types';
 
 const ROOT_TEST_ID = 'textfield';
 const noop = () => undefined;
