+++ packages/design-system-react/src/components/TextField/TextField.types.ts
@@ -3,6 +3,32 @@ import type { ComponentPropsWithoutRef, Ref } from 'react';
 
 import type { InputProps } from '../Input/Input.types';
 
+/**
+ * TextField size variants (web-only, ADR-0003).
+ *
+ * The React Native `TextField` ships a fixed height and intentionally does not
+ * expose a size prop, so this const object lives in the web package only.
+ */
+export const TextFieldSize = {
+  Sm: 'sm',
+  Md: 'md',
+  Lg: 'lg',
+} as const;
+export type TextFieldSize = (typeof TextFieldSize)[keyof typeof TextFieldSize];
+
+/**
+ * Native input `type` attribute values supported by the web TextField
+ * (ADR-0003). React Native consumers use `inputProps.keyboardType` /
+ * `inputProps.secureTextEntry` instead, so this is web-only.
+ */
+export const TextFieldType = {
+  Text: 'text',
+  Number: 'number',
+  Password: 'password',
+  Search: 'search',
+} as const;
+export type TextFieldType = (typeof TextFieldType)[keyof typeof TextFieldType];
+
 type TextFieldInputProps = Omit<
   InputProps,
   | 'autoFocus'
@@ -26,6 +52,40 @@ export type TextFieldProps = Omit<
   'onBlur' | 'onChange' | 'onFocus' | 'onClick'
 > &
   TextFieldPropsShared & {
+    /**
+     * The size of the text field. Controls the height.
+     *
+     * @default TextFieldSize.Md
+     */
+    size?: TextFieldSize;
+    /**
+     * The native input `type` attribute.
+     *
+     * @default TextFieldType.Text
+     */
+    type?: TextFieldType;
+    /**
+     * If true, truncates overflowing input text with an ellipsis.
+     *
+     * @default true
+     */
+    truncate?: boolean;
+    /**
+     * Max number of characters to allow.
+     */
+    maxLength?: number;
+    /**
+     * Name attribute of the inner `input` element.
+     */
+    name?: string;
+    /**
+     * `id` of the inner `input` element.
+     */
+    id?: string;
+    /**
+     * If true, the inner input is marked as required.
+     */
+    required?: boolean;
     /**
      * Additional props for the inner `Input`. Do not pass `value`,
      * `placeholder`, `isDisabled`, `isReadOnly`, `onFocus`, `onBlur`,
