+++ packages/design-system-react/src/components/TextField/index.ts
@@ -1,3 +1,3 @@
-export { TextFieldSize, TextFieldType } from '@metamask/design-system-shared';
 export { TextField } from './TextField';
+export { TextFieldSize, TextFieldType } from './TextField.types';
 export type { TextFieldProps } from './TextField.types';
