+++ packages/design-system-shared/src/types/TextField/index.ts
@@ -1,5 +1,1 @@
-export {
-  TextFieldSize,
-  TextFieldType,
-  type TextFieldPropsShared,
-} from './TextField.types';
+export { type TextFieldPropsShared } from './TextField.types';
